+++ BOString/NSAttributedString+BOString.h
@@ -29,6 +29,15 @@
  *  `NSAttributedString`. In case of conflicts initial attributes are 
  *  re-written.
  */
+- (NSAttributedString *)bos_makeString:(void(^)(BOStringMaker *make))block;
+
+#ifdef BOS_SHORTHAND
+
+/**
+ *  Shorthand method for bos_makeString:.
+ */
 - (NSAttributedString *)makeString:(void(^)(BOStringMaker *make))block;
+
+#endif // BOS_SHORTHAND
 
 @end